Fixed the location of mode table.
It feeled weird we have to compute it from `xwia` location, which may be not present at all (but we used it anyway). Tracing has shown that the offset computed using the formula matches the offset in `wmta` field, so it is likely that that field should be used instead.